fix: address 2 unresolved PR comments + fix stale test parsers
- P1: Enforce same-user ownership check on session_message handler,
  especially for deliverAs:input which starts new agent turns
- P2: Make tree expand toggle keyboard-focusable (tabIndex 0 vs -1)
- Extract trigger parsers into trigger-parsers.ts so tests import the
  real implementation instead of a stale copy with different logic
